switch mojo tests to ITs, so they don’t fail jenkins Streams Trunk build
