+++ streams-plugins/streams-plugin-elasticsearch/src/site/markdown/index.md
@@ -11,7 +11,7 @@ Output will be placed in target/generated-resources/elasticsearch by default
 
 Run within a module containing a src/main/jsonschema directory
 
-    mvn org.apache.streams.plugins:streams-plugin-elasticsearch:0.3-incubating-SNAPSHOT:elasticsearch
+    mvn org.apache.streams.plugins:streams-plugin-elasticsearch:0.4-incubating:elasticsearch
 
 [streams-plugin-elasticsearch/pom.xml](streams-plugin-elasticsearch/pom.xml "streams-plugin-elasticsearch/pom.xml")
