+++ streams-plugins/streams-plugin-hbase/src/site/markdown/index.md
@@ -13,7 +13,7 @@ Output will be placed in target/generated-resources/hbase by default
 
 Run within a module containing a src/main/jsonschema directory
 
-    mvn org.apache.streams.plugins:streams-plugin-hbase:0.3-incubating-SNAPSHOT:hbase
+    mvn org.apache.streams.plugins:streams-plugin-hbase:0.4-incubating:hbase
 
 ##### SDK
