+++ streams-plugins/streams-plugin-hbase/src/test/java/org/apache/streams/plugins/test/StreamsHbaseResourceGeneratorMojoIT.java
@@ -16,9 +16,11 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.streams.plugins.elasticsearch.test;
+
+package org.apache.streams.plugins.test;
 
 import com.google.common.collect.Lists;
+import com.google.common.io.Files;
 import junit.framework.TestCase;
 import org.apache.maven.it.Verifier;
 import org.apache.maven.it.util.ResourceExtractor;
@@ -28,14 +30,17 @@ import org.slf4j.LoggerFactory;
 
 import java.io.File;
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.List;
+
+import static org.apache.streams.plugins.test.StreamsHbaseResourceGeneratorTest.txtFilter;
 
 /**
  * Tests that streams-plugin-hive running via maven generates hql resources
  */
-public class StreamsElasticsearchResourceGeneratorMojoTest extends TestCase {
+public class StreamsHbaseResourceGeneratorMojoIT extends TestCase {
 
-    private final static Logger LOGGER = LoggerFactory.getLogger(StreamsElasticsearchResourceGeneratorMojoTest.class);
+    private final static Logger LOGGER = LoggerFactory.getLogger(StreamsHbaseResourceGeneratorMojoIT.class);
 
     protected void setUp() throws Exception
     {
@@ -45,9 +50,9 @@ public class StreamsElasticsearchResourceGeneratorMojoTest extends TestCase {
 
 
     @Test
-    public void testStreamsElasticsearchResourceGeneratorMojo() throws Exception {
+    public void testStreamsHbaseResourceGeneratorMojo() throws Exception {
 
-        File testDir = ResourceExtractor.simpleExtractResources( getClass(), "/streams-plugin-elasticsearch" );
+        File testDir = ResourceExtractor.simpleExtractResources( getClass(), "/streams-plugin-hbase" );
 
         Verifier verifier;
 
@@ -64,5 +69,15 @@ public class StreamsElasticsearchResourceGeneratorMojoTest extends TestCase {
 
         verifier.resetStreams();
 
+        File testOutput = new File(testDir.getAbsolutePath() + "/target/generated-resources/hbase-mojo");
+
+        assert( testOutput != null );
+        assert( testOutput.exists() == true );
+        assert( testOutput.isDirectory() == true );
+
+        Iterable<File> outputIterator = Files.fileTreeTraverser().breadthFirstTraversal(testOutput)
+                .filter(txtFilter);
+        Collection<File> outputCollection = Lists.newArrayList(outputIterator);
+        assert( outputCollection.size() == 133 );
     }
 }
