+++ streams-plugins/streams-plugin-pig/src/test/java/org/apache/streams/plugins/pig/test/StreamsPigResourceGeneratorMojoIT.java
@@ -16,12 +16,12 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.streams.plugins.cassandra.test;
+
+package org.apache.streams.plugins.pig.test;
 
 import com.google.common.collect.Lists;
 import com.google.common.io.Files;
 import junit.framework.TestCase;
-import org.apache.commons.lang3.StringUtils;
 import org.apache.maven.it.Verifier;
 import org.apache.maven.it.util.ResourceExtractor;
 import org.junit.Test;
@@ -29,20 +29,18 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.File;
-import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
 
-import static org.apache.streams.plugins.cassandra.test.StreamsCassandraResourceGeneratorTest.cqlFilter;
+import static org.apache.streams.plugins.pig.test.StreamsPigResourceGeneratorTest.pigFilter;
 
 /**
  * Tests that streams-plugin-hive running via maven generates hql resources
  */
-public class StreamsCassandraResourceGeneratorMojoTest extends TestCase {
+public class StreamsPigResourceGeneratorMojoIT extends TestCase {
 
-    private final static Logger LOGGER = LoggerFactory.getLogger(StreamsCassandraResourceGeneratorMojoTest.class);
+    private final static Logger LOGGER = LoggerFactory.getLogger(StreamsPigResourceGeneratorMojoIT.class);
 
     protected void setUp() throws Exception
     {
@@ -50,10 +48,11 @@ public class StreamsCassandraResourceGeneratorMojoTest extends TestCase {
         super.setUp();
     }
 
-    @Test
-    public void testStreamsCassandraResourceGeneratorMojo() throws Exception {
 
-        File testDir = ResourceExtractor.simpleExtractResources( getClass(), "/streams-plugin-cassandra" );
+    @Test
+    public void testStreamsPigResourceGeneratorMojo() throws Exception {
+
+        File testDir = ResourceExtractor.simpleExtractResources( getClass(), "/streams-plugin-pig" );
 
         Verifier verifier;
 
@@ -70,27 +69,15 @@ public class StreamsCassandraResourceGeneratorMojoTest extends TestCase {
 
         verifier.resetStreams();
 
-        Path testOutputPath = Paths.get(testDir.getAbsolutePath()).resolve("target/generated-resources/test-mojo");
-
-        File testOutput = testOutputPath.toFile();
+        File testOutput = new File(testDir.getAbsolutePath() + "/target/generated-resources/pig-mojo");
 
         assert( testOutput != null );
         assert( testOutput.exists() == true );
         assert( testOutput.isDirectory() == true );
 
         Iterable<File> outputIterator = Files.fileTreeTraverser().breadthFirstTraversal(testOutput)
-                .filter(cqlFilter);
+                .filter(pigFilter);
         Collection<File> outputCollection = Lists.newArrayList(outputIterator);
-        assert( outputCollection.size() == 1 );
-
-        Path path = testOutputPath.resolve("types.cql");
-
-        assert( path.toFile().exists() );
-
-        String typesCqlBytes = new String(
-                java.nio.file.Files.readAllBytes(path));
-
-        assert( StringUtils.countMatches(typesCqlBytes, "CREATE TYPE") == 133 );
-
+        assert( outputCollection.size() == 133 );
     }
 }
